Generalise clear/class to details summary/specifics
Instead of a specific Clear entry which adds a Clear: header and a
specific Class entry, generalise these to Details with Summary (in bold)
and Specifics (normal).

This gives the user more control over what appears, for example listing
critical damage rules on the card.

Also fixes the class position on items which are more than one unit high.

Fixes #14.